Rename workflowChoice constants
